refactor: rename StateTransitionUI to TaskStateTransitionUI and update related components for task state management
