+++ src/main/kotlin/presentation/task/TaskStateTransitionUI.kt
@@ -1,17 +1,17 @@
-package org.baghdad.presentation
+package org.baghdad.presentation.task
 
 import kotlinx.coroutines.runBlocking
 import org.baghdad.logic.manager.SessionManager
 import org.baghdad.logic.model.entities.StateEntity
 import org.baghdad.logic.model.entities.TaskEntity
 import org.baghdad.logic.model.exceptions.NotFoundException
-import org.baghdad.logic.usecase.StateTransitionUseCase
+import org.baghdad.logic.usecase.task.TaskStateTransitionUseCase
 import org.baghdad.presentation.input.Reader
 import org.baghdad.presentation.output.Viewer
-import java.util.*
+import java.util.UUID
 
-class StateTransitionUI(
-    private val useCase: StateTransitionUseCase,
+class TaskStateTransitionUI(
+    private val useCase: TaskStateTransitionUseCase,
     private val session: SessionManager,
     private val viewer: Viewer,
     private val reader: Reader,
