+++ src/test/kotlin/logic/usecase/task/TaskStateTransitionUseCaseTest.kt
@@ -1,7 +1,10 @@
-package logic.usecase
+package logic.usecase.task
 
-
-import io.mockk.*
+import io.mockk.Runs
+import io.mockk.coEvery
+import io.mockk.coVerify
+import io.mockk.just
+import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.baghdad.logic.model.entities.StateEntity
 import org.baghdad.logic.model.entities.TaskEntity
@@ -12,22 +15,21 @@ import org.baghdad.logic.repositories.AuditRepository
 import org.baghdad.logic.repositories.ProjectStatesRepository
 import org.baghdad.logic.repositories.TaskRepository
 import org.baghdad.logic.repositories.UserRepository
-import org.baghdad.logic.usecase.StateTransitionUseCase
+import org.baghdad.logic.usecase.task.TaskStateTransitionUseCase
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
-import java.util.*
+import java.util.UUID
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-
-class StateTransitionUseCaseTest {
+class TaskStateTransitionUseCaseTest {
 
     private val taskRepository = mockk<TaskRepository>()
     private val projectStatesRepository = mockk<ProjectStatesRepository>()
     private val userRepository = mockk<UserRepository>() // NEW
     private val auditRepository = mockk<AuditRepository>()
-    private lateinit var service: StateTransitionUseCase
+    private lateinit var service: TaskStateTransitionUseCase
 
     private lateinit var task: TaskEntity
     private lateinit var oldState: StateEntity
@@ -36,7 +38,7 @@ class StateTransitionUseCaseTest {
 
     @BeforeEach
     fun setup() {
-        service = StateTransitionUseCase(taskRepository, projectStatesRepository, userRepository, auditRepository)
+        service = TaskStateTransitionUseCase(taskRepository, projectStatesRepository, userRepository, auditRepository)
 
         val projectId = UUID.randomUUID()
         val oldStateId = UUID.randomUUID()
